Allow Activity to define theme for ComposeView
