+++ collect_app/build.gradle
@@ -4,6 +4,7 @@ plugins {
     alias(libs.plugins.kotlinParcelize)
     alias(libs.plugins.ossLicenses)
     alias(libs.plugins.safeargs)
+    alias(libs.plugins.composeCompiler)
 }
 apply from: '../config/quality.gradle'
 
@@ -76,6 +77,7 @@ android {
     buildFeatures {
         viewBinding true
         buildConfig true
+        compose true
     }
 
     defaultConfig {
@@ -346,6 +348,11 @@ dependencies {
 
     implementation libs.splashscreen
     implementation libs.androidxNavigationFragmentKtx
+
+    var composeBom = platform("androidx.compose:compose-bom:2025.05.00")
+    implementation(composeBom)
+    implementation("androidx.compose.material3:material3")
+    implementation("androidx.constraintlayout:constraintlayout-compose:1.0.1")
 
     testImplementation project(':forms-test')
