+++ collect_app/src/main/java/org/odk/collect/android/activities/ScannerWithFlashlightActivity.kt
@@ -1,0 +1,38 @@
+/*
+ * Copyright (C) 2017 Shobhit
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package org.odk.collect.android.activities
+
+import android.os.Bundle
+import android.view.Window
+import android.view.WindowManager
+import org.odk.collect.android.R
+import org.odk.collect.android.application.CollectComposeThemeProvider
+import org.odk.collect.strings.localization.LocalizedActivity
+
+/**
+ * Custom Scannner Activity extending from Activity to display a custom layout form scanner view.
+ */
+class ScannerWithFlashlightActivity : LocalizedActivity(), CollectComposeThemeProvider {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        requestWindowFeature(Window.FEATURE_NO_TITLE)
+        window.setFlags(
+            WindowManager.LayoutParams.FLAG_FULLSCREEN,
+            WindowManager.LayoutParams.FLAG_FULLSCREEN
+        )
+
+        setContentView(R.layout.activity_custom_scanner)
+    }
+}
